test: isolate e2e tests from root project
Will ease the support of old PHP version with downgrade and keep developing using new PHP versions
