+++ composer.json
@@ -22,8 +22,7 @@
         "phpstan/phpstan": "^1.11",
         "phpstan/phpstan-phpunit": "^1.4",
         "phpstan/phpstan-strict-rules": "^1.6",
-        "phpstan/phpstan-symfony": "^1.4",
-        "symfony/process": "^6.4"
+        "phpstan/phpstan-symfony": "^1.4"
     },
     "autoload": {
         "psr-4": {
@@ -32,13 +31,11 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Buildotter\\Tests\\MakerStandalone\\": "tests",
-            "Buildotter\\E2e\\MakerStandalone\\": "e2e"
+            "Buildotter\\Tests\\MakerStandalone\\": "tests"
         },
         "files": [
             "tests/functions.php"
-        ],
-        "exclude-from-classmap": ["/e2e/cases/"]
+        ]
     },
     "config": {
         "platform": {
