+++ e2e/.phive/phars.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phive xmlns="https://phar.io/phive">
+  <phar name="phpunit" version="^10.5.27" installed="10.5.27" location="./tools/phpunit.phar" copy="true"/>
+</phive>
