+++ e2e/phpunit.xml.dist
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/10.5/phpunit.xsd"
-         bootstrap="../vendor/autoload.php"
+         bootstrap="./vendor/autoload.php"
          cacheDirectory=".phpunit.cache"
          colors="true"
          executionOrder="random"
