+++ phpstan.dist.neon
@@ -16,9 +16,6 @@ parameters:
 	paths:
 		- src/
 		- tests/
-		- e2e/
-	excludePaths:
-		- e2e/cases/*
 	symfony:
 		consoleApplicationLoader: tests/console-app.php
 	ignoreErrors:
